// src/lib/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  cooking: {
    Tables: {
      ingredient_catalog: {
        Row: {
          created_at: string
          id: string
          name: string
          user_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          name: string
          user_id: string
        }
        Update: {
          created_at?: string
          id?: string
          name?: string
          user_id?: string
        }
        Relationships: []
      }
      ingredients: {
        Row: {
          catalog_ingredient_id: string
          id: string
          quantity: number
          recipe_id: string
          unit: string
        }
        Insert: {
          catalog_ingredient_id: string
          id?: string
          quantity: number
          recipe_id: string
          unit: string
        }
        Update: {
          catalog_ingredient_id?: string
          id?: string
          quantity?: number
          recipe_id?: string
          unit?: string
        }
        Relationships: [
          {
            foreignKeyName: "ingredients_catalog_ingredient_id_fkey"
            columns: ["catalog_ingredient_id"]
            isOneToOne: false
            referencedRelation: "ingredient_catalog"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "ingredients_recipe_id_fkey"
            columns: ["recipe_id"]
            isOneToOne: false
            referencedRelation: "recipes"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      recipes: {
        Row: {
          created_at: string
          description: string | null
          id: string
          name: string
          servings: number
          user_id: string
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: string
          name: string
          servings?: number
          user_id: string
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: string
          name?: string
          servings?: number
          user_id?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}